refactor toml config struct
