+++ src/theme.rs
@@ -2,6 +2,6 @@ mod helper;
 pub mod installed;
 pub mod online;
 pub mod saved;
-
+mod toml_config;
 pub use helper::create_theme_id;
 pub use helper::ThemeId;
